split client methods into smaller modules
